requirements.txt
Django==5.2.1
PyMySql==1.1.1
bcrypt==4.3.0
pip==25.0.1
transbank-sdk==6.0.0
djangorestframework==3.16.0